Modify start_raw() to get raw EMG and add new functionalities
Although the original code is supposed to read the raw EMG signals measured by the sensor pods of the Myo, it actually reads rectified and smoothed EMG signals, a measure of their amplitude. The code has been modified to get the real raw signals, by suscribing to the raw EMG notification characteristics by writing to and reading from the corresponding handles listed on the Thalmic Labs BLE specification.
A new function to change the color of the logo and bar LEDs has been added.
The program is now able to get the battery level in the same manner that it gets the EMG, IMU or classifier data. By suscribing to the battery characteristic, the Myo sends the battery level every time its value changes.